Maximum confirmation changes, some fixes to getBtcTransactionConfirmations, unit tests
- Method getBtcTransactionConfirmations now only allows getting confirmations for transactions in blocks with at most 4320 confirmations
- Unified way in which we get the best chain height within getBtcTransactionConfirmations
- Updated unit tests
- Bridge parameter parsing is using bytes now instead of string
- Fixed method signature
- Passing in a MerkleBranch instance to BridgeSupport instead of individual parameters
- Fixed possible error codes, grouping those that represent inconsistent internal states
- New helper MerkleTreeUtils, with method combineLeftRight
- Unit tested MerkleTreeUtils
- Used MerkleTreeUtils.combineLeftRight both in MerkleBranch and in Rskip92MerkleProofValidator
- Fixed and completed BridgeSupport.getBtcTransactionConfirmations and Bridge.getBtcTransactionConfirmations unit tests
- Added construction validation unit tests for MerkleBranch

+++ rskj-core/src/main/java/co/rsk/peg/BridgeMethods.java
@@ -163,7 +163,7 @@ public enum BridgeMethods {
             CallTransaction.Function.fromSignature(
                     "getBtcTransactionConfirmations",
                     new String[]{"bytes32", "bytes32", "uint256", "bytes32[]"},
-                    new String[]{"int64"}
+                    new String[]{"int256"}
             ),
             22000L, //TODO ESTIMATE GAS
             (BridgeMethodExecutorTyped) Bridge::getBtcTransactionConfirmations,
